Cham cong: tinh CONG theo 2 CA (sang 8:00-11:45, chieu 13:30-17:00)
- Attendance.work_credit: ca sang + ca chieu, moi ca 0.5 cong. Ca ngay=1, nua ngay=0.5.
  Thieu gio ra -> suy theo gio vao; co cham vao ma khong roi ca nao -> toi thieu 0.5.
- Summary "Ngay cong" (present_days) + payroll working_days GIO cong theo ca (so thap phan)
  thay vi dem so ngay -> nguoi lam nua ngay tra dung nua cong / nua luong.
- Cau hinh moc ca o config (MORNING_END_MIN / AFTERNOON_START_MIN). Khong them cot DB.

+++ backend/app/schemas.py
@@ -493,7 +493,7 @@ class AttendanceSummary(BaseModel):
     """Tổng hợp chấm công theo người trong 1 kỳ (cho màn quản lý)."""
     user_id: int
     full_name: str
-    present_days: int = 0
+    present_days: float = 0.0   # NGÀY CÔNG theo ca: cả ngày = 1, nửa ngày = 0.5
     late_days: int = 0
     total_hours: float = 0.0
